Create Workshop Day 2- Alternatives
Dplyr is your friend when it comes to data manipulation! It provides easy to use tools to get your dataset where you need it to be before you run your analyses.

However, learning some of these functions in basic R is helpful as well. First, it gives you the opportunity to practice indexing. Second, some secure servers might not allow you to load tidyverse. Third, if you aren't sure about something in dplyr or struggling to get the code to work, you have some familiarity with an alternative.